Split SvgGaussianBlur
